requirements.txt
imagehash
pillow 
numpy
